printf: fix padding count for widths of 256 or more
- printf("%300d", 1) output only 43 spaces
- 299 % 256 = 43 because the padding count was calculated modulo 256

+++ clibs/printf.c
@@ -25,7 +25,7 @@ static  void  putchar_to_console(uint8_t c)
   out_count++;
 }
 
-static void putchar_rep(uint8_t c, uint8_t  n)
+static void putchar_rep(uint8_t c, int n)
 {
   while (n) {
     putchar_p(c);
